+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `glsl-lang`: Implement PartialOrd and Hash for Identifier and TypeName
 - `lang-util`: Add NodeSpan::start() and NodeSpan::end()
 - `lang-util`: Generate display impl for derive(Token)
+- `lang-util`: Add lang_util(as = "...") for mapping to lalrpop parser tokens
 
 ### Changed
 - PartialEq for Node now compares the contents of the node by default